Major debugging work! Most of the features are not implemented, as most of the work has been on the architecture and the API. There were some heavy changes in the Debugging namespace, and more will probably be coming as I start implementing things. Since debugging was only partially implemented before and semi-torn out, not many changes were made to the plugins, except for statistics. I've added a sort-of complete performance-counter like implementation. I'm not sure I'm happy with it yet, so it may change again.
A new project was added: the RemoteDebugger. This is a separate executable application. The log view was removed from the emulator - replacing it is a dummy logger that just writes to the debug console - the real logger is now in the debugger. The debugger uses .NET Remoting, and with a little work could be binary compatible with Mono (as most of the types serialized are my own - only a few collections would need tweaking) - that means you could use the Windows debugger to debug a Linux version of the emulator, etc! More realistically, it will allow for robust local debugging (the debugger will stay up even if the emulator crashes, stay active even if the emulator deadlocks, or more importantly stay active if the emulator is being debugged in VS).
Currently the remote debugger has the log viewer and a fancy connection dialog. Oh, and a pretty new icon.
As for connections, I auto connect to localhost for now. If anyone is actually using remote debugging, I can actually save the last connection/etc. If you hit Start Debugging, the app will keep trying to connect to the emulator - this allows the debugger to start and wait until the emulator starts. If you have the emulator open first, you can hit 'Start with Debugging' to have the emulator wait until a debugger connects.
I don't current have anything in for waiting on a commandline start for the emulator. I'll add that in a bit.

Lots of work to do here - next will probably be a very eary statistics viewer for the debugger (and tweaking of the stats system until I'm happy), then some work on the CPU to start supporting debug control.

NOTE: I don't think my debug settings check in (on purpose), so if you are doing dev, you should change the solution to start multiple projects on debug. You can do this by right clicking the solution -> Properties -> select 'Multiple startup projects', set both RemoteDebugger and Player to Start, and make sure RemoteDebugger is ordered above Player.

+++ trunk/Noxa.Emulation.Psp.Video.OpenGL/OglDriver_NativeInterface.cpp
@@ -56,7 +56,7 @@ bool _vsyncWaiting;
 int64 _lastVsync;
 
 // Statistics
-extern uint _vcount;
+extern uint64 _vcount;
 
 void __break()
 {
